MAINT - Error constants and warnings - added markers for all error numbers; - reviewed and documented errors (in spreadsheet for wiki); - removed unthrown exceptions in method definitions; - removed unused variables and code; - changed some private method names to be more descriptive of function; - explicit types for generic declarations; - removed associated and unnecessary casts; and - added missing `@Override` annotations.

+++ jhove-modules/src/main/java/edu/harvard/hul/ois/jhove/module/pdf/PdfIndirectObj.java
@@ -41,7 +41,7 @@ public class PdfIndirectObj extends PdfObject
 	        return _cachedObject;
         }
         long key = ((long) _objNumber << 32) +
-            ((long) _genNumber & 0XFFFFFFFFL);
+            (_genNumber & 0XFFFFFFFFL);
         _cachedObject = _objectMap.get (new Long (key));
         return _cachedObject;
     }
